agent/pyproject.toml
[project]
name = "mock-interview-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=0.12",
    "livekit-plugins-deepgram>=0.9",
    "livekit-plugins-openai>=0.10",
    "livekit-plugins-elevenlabs>=0.9",
    "livekit-plugins-silero>=0.7",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
